+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## Unreleased Changes
 
+## v0.2.0
+
 * Added AT25SF128A support
 * Ensured all bootloaders are built on a `cargo build`
+* Enable building without GCC by providing precompiled binaries. Use `--feature=assemble` to opt out.
+* Added CI using `--feature=assemble` to verify latest boot2 source matches bootloader binary blobs
 
 ## v0.1.2